Training - refactor bot.py file strutcure
